Cambios en validar

+++ validar.py
@@ -137,7 +137,7 @@ class validador():
         try:
             float(trama[2])
             float(trama[3])
-            int(trama[4])
+            float(trama[4])
             bool(trama[5])
             int(trama[6])
             bool(trama[7])
